Detect a 'swiping' hand movement
Added code in 'loop' to read input from all IR sensors as either LOW or HIGH. Then, the array storing the sensor states will be traversed to check whether or not all are activated. If all sensors are activated (HIGH), a swipe has been detected. Otherwise, if any one of the sensors is not activated (LOW), then a swipe is not detected. In the latter case, there is no point in checking the remaining sensors, so simply exit the loop.