Adding 'taskOnSwipe' function
Instead of having the logic for the specific colour to turn on in 'loop', move this to a separate function for simplicity and abstraction of code. The 'taskOnSwipe' function will be passed the 'mySwipe' parameter which contains the current number of swipes. The function will use this parameter against case statements in a 'switch' block to determine which colour to change the light to. The default case is to turn the light off.